requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
azure-search-documents
azure-identity
openai
PyPDF2
python-jose
requests
